rearranged file structure to move components out of pages folder
