+++ ace-your-day/src/App.js
@@ -3,7 +3,6 @@ import './App.css';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import Home from './components/pages/home'
 import Calendar from './components/pages/calendar'
-import Des from './components/pages/eventDescription'
 
 
 function App() {
@@ -13,7 +12,6 @@ function App() {
         <Routes>
           <Route path='/' element={<Home/>}></Route>
           <Route path='/calendar' element={<Calendar/>}></Route>
-          <Route path='/des' element={<Des/>}></Route>
         </Routes>
       </Router>
